feat: document yserde
